ID 데이터 타입 변경: Integer -> Long

+++ src/main/java/com/delta/delta/controller/UserController.java
@@ -13,7 +13,7 @@ public class UserController {
     UserService userService;
 
     @GetMapping("/{userId}")
-    public User getUser(@PathVariable Integer userId) {
+    public User getUser(@PathVariable Long userId) {
         return userService.getUser(userId);
     }
 
@@ -23,32 +23,32 @@ public class UserController {
     }
 
     @PatchMapping("/{userId}")
-    public User updateUser(@PathVariable Integer userId, @RequestBody User user) {
+    public User updateUser(@PathVariable Long userId, @RequestBody User user) {
         return userService.updateUser(userId, user);
     }
 
     @DeleteMapping("/{userId}")
-    public void deleteUser(@PathVariable Integer userId) {
+    public void deleteUser(@PathVariable Long userId) {
         userService.deleteUser(userId);
     }
 
     @PostMapping("/addFollower/{userId}/{followerId}")
-    public void addFollower(@PathVariable Integer userId, @PathVariable Integer followerId) {
+    public void addFollower(@PathVariable Long userId, @PathVariable Long followerId) {
         userService.addFollower(userId, followerId);
     }
 
     @PostMapping("/deleteFollower/{userId}/{followerId}")
-    public void deleteFollower(@PathVariable Integer userId, @PathVariable Integer followerId) {
+    public void deleteFollower(@PathVariable Long userId, @PathVariable Long followerId) {
         userService.deleteFollower(userId, followerId);
     }
 
     @PostMapping("/addFollowing/{userId}/{followingId}")
-    public void addFollowing(@PathVariable Integer userId, @PathVariable Integer followingId) {
+    public void addFollowing(@PathVariable Long userId, @PathVariable Long followingId) {
         userService.addFollowing(userId, followingId);
     }
 
     @PostMapping("/deleteFollowing/{userId}/{followingId}")
-    public void deleteFollowing(@PathVariable Integer userId, @PathVariable Integer followingId) {
+    public void deleteFollowing(@PathVariable Long userId, @PathVariable Long followingId) {
         userService.deleteFollowing(userId, followingId);
     }
 }
